// frontend/src/client/types.gen.ts
export type InvoiceCreate = {
    invoice_number: string;
    invoice_code: string;
    amount: number;
    invoice_date?: (string | null);
    seller?: (string | null);
    buyer?: (string | null);
    tax_amount?: (number | null);
    remark?: (string | null);
    contract_id: string;
};